backend/authentication/serializers.py
# -*- coding: utf-8 -*-
from rest_framework import serializers

from project.models import User


class UserSerializer(serializers.ModelSerializer):
    username = serializers.CharField(max_length=255)
    password = serializers.CharField(max_length=128)

    class Meta:
        model = User
        fields = ('username', 'password')

    def create(self, validated_data):
        return User.objects.create_user(**validated_data)
